Fixed a bug where the cd timer was not calculated correctly, it should now be working as intend

+++ Hirathiel/Hirathiel/Game.hpp
@@ -35,5 +35,5 @@ private:
 	Window* window;
 	Player* player;
 	MobList* mobs;
-	int cd = 1;
+	float cd = 1;
 };
